+++ packages/@tinacms/toolkit/src/hooks/use-local-storage.ts
@@ -1,0 +1,36 @@
+/**
+Copyright 2021 Forestry.io Holdings, Inc.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import * as React from 'react'
+
+export function useLocalStorage(key, initialValue) {
+  const [storedValue, setStoredValue] = React.useState(initialValue)
+  React.useEffect(() => {
+    const valueFromStorage = window.localStorage.getItem(key)
+    if (valueFromStorage != null) {
+      setStoredValue(JSON.parse(valueFromStorage))
+    }
+  }, [key])
+
+  const setValue = (value) => {
+    try {
+      const valueToStore =
+        value instanceof Function ? value(storedValue) : value
+      setStoredValue(valueToStore)
+      localStorage.setItem(key, JSON.stringify(valueToStore))
+    } catch (error) {
+      console.log(error)
+    }
+  }
+  return [storedValue, setValue]
+}