+++ packages/react-tinacms-editor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 0.52.7
+
+### Patch Changes
+
+- Updated dependencies [df3030990]
+  - @tinacms/toolkit@0.56.0
+
+## 0.52.6
+
+### Patch Changes
+
+- Updated dependencies [60729f60c]
+  - @tinacms/toolkit@0.55.4
+
 ## 0.52.5
 
 ### Patch Changes